AiService/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
redis
qdrant-client
numpy
scikit-learn
beautifulsoup4
openai